requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
sqlmodel==0.0.22
apscheduler==3.10.4
google-api-python-client==2.161.0
google-auth==2.38.0
google-generativeai==0.8.4
ffmpeg-python==0.2.0
httpx==0.28.1
python-multipart==0.0.20
aiofiles==24.1.0
tenacity==9.0.0
